pyproject.toml
[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"

[project]
name = "ivs-forecast"
version = "0.1.0"
description = "Chronology-safe next-day implied volatility surface forecasting framework."
readme = "README.md"
requires-python = ">=3.14.3,<3.15"
dependencies = [
  "numpy",
  "polars",
  "pyarrow",
  "pydantic",
  "pyyaml",
  "rich",
  "scipy",
  "torch",
  "typer",
  "xgboost",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[project.scripts]
ivs-forecast = "ivs_forecast.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["B008", "E501"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
